refactor: ReservationTable -> Reservations로 이름 변경
