+++ src/main/java/roomescape/reservation/ReservationController.java
@@ -6,7 +6,7 @@ import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationId;
 import roomescape.reservation.dto.CreateReservationBody;
 import roomescape.reservation.dto.ReservationDto;
-import roomescape.reservation.dto.ReservationTable;
+import roomescape.reservation.domain.Reservations;
 
 import java.net.URI;
 import java.util.List;
@@ -15,12 +15,12 @@ import java.util.concurrent.atomic.AtomicLong;
 @RestController
 @RequestMapping("/reservations")
 public class ReservationController {
-	private final ReservationTable reservationTable = new ReservationTable();
+	private final Reservations reservations = new Reservations();
 	private final AtomicLong nextId = new AtomicLong(0);
 	
 	@GetMapping
 	public ResponseEntity<List<ReservationDto>> getReservations() {
-		List<ReservationDto> reservations = reservationTable.getAll().stream()
+		List<ReservationDto> reservations = this.reservations.getAll().stream()
 				.map(ReservationDto::new)
 				.toList();
 		return ResponseEntity.ok(reservations);
@@ -30,7 +30,7 @@ public class ReservationController {
 	public ResponseEntity<ReservationDto> createReservation(@RequestBody CreateReservationBody createReservationBody) {
 		ReservationId id = new ReservationId(nextId.incrementAndGet());
 		Reservation reservation = createReservationBody.createEntity(id);
-		reservationTable.add(reservation);
+		reservations.add(reservation);
 		
 		return ResponseEntity
 				.created(URI.create("/reservations/" + reservation.getId()))
@@ -40,7 +40,7 @@ public class ReservationController {
 	@DeleteMapping("{id}")
 	public ResponseEntity<Void> deleteReservation(@PathVariable long id) {
 		ReservationId reservationId = new ReservationId(id);
-		reservationTable.remove(reservationId);
+		reservations.remove(reservationId);
 		
 		return ResponseEntity.noContent().build();
 	}
